pyproject.toml
[project]
name = "apa"
version = "0.1.0"
description = "Email-triggered agentic process automation powered by Claude and Playwright"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "a2a-sdk>=0.3.25",
    "claude-agent-sdk>=0.1.0",
    "fastapi>=0.115.0",
    "imapclient>=3.1.0",
    "pika>=1.3.0",
    "python-dotenv>=1.2.2",
    "uvicorn[standard]>=0.34.0",
]

[project.scripts]
apa = "apa.__main__:main"

[dependency-groups]
dev = [
    "pyright>=1.1.408",
    "ruff>=0.15.9",
]

[tool.pyright]
venvPath = "."
venv = ".venv"

[tool.ruff]
target-version = "py312"
line-length = 99

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
